Updating structure of sitemap parser
